refactor: extract createTaskStatusQuery to check-helpers.ts
Deduplicates the task-store fetch/envelope-unwrap boilerplate against
createPrRecordQuery, matching check-helpers.ts's role as the shared
home for WL-2.2 candidate-provider helpers. Also tightens the return
type from string to TaskStatus.
