+++ agent/src/check-deploy.ts
@@ -23,6 +23,7 @@
 import {
   candidateId,
   createPrRecordQuery,
+  createTaskStatusQuery,
   getCurrentUser,
   isCleanApproveBody,
   readAllowSelfReview,
@@ -30,6 +31,7 @@ import {
   resolveWorkspacePath,
   splitOrgRepo,
 } from "./check-helpers.ts";
+import type { TaskStatus } from "./check-helpers.ts";
 import type { WorkPrCandidate } from "./work-selector.ts";
 
 // ─── Types ────────────────────────────────────────────────────────────────────
@@ -87,7 +89,7 @@ export interface CheckDeployDeps {
   queryPrRecord?: (repo: string, prNumber: number) => Promise<PrRecord | null>;
   // Task status lookup for the linked task (if any). Returns the task's status
   // or null if no task is found or lookup fails. Used to exclude blocked PRs.
-  queryTaskStatus?: (repo: string, prNumber: number) => Promise<string | null>;
+  queryTaskStatus?: (repo: string, prNumber: number) => Promise<TaskStatus | null>;
 }
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
@@ -331,53 +333,5 @@ export async function buildProductionDeps(opts: {
     isBundleComplete: undefined,
     queryPrRecord: createPrRecordQuery<PrRecord>({ fetchFn: opts.fetchFn }),
     queryTaskStatus: createTaskStatusQuery({ fetchFn: opts.fetchFn }),
-  };
-}
-
-/**
- * Build a `(repo, prNumber) => status | null` query function against the
- * task-store `/tasks` endpoint with `?repo=&pr=` filters. Returns null on
- * missing config, a non-ok response, no matching task, or any fetch error —
- * a missing task-store task record must not throw.
- *
- * Accepts an optional `fetchFn` for dependency injection in tests, matching
- * createPrRecordQuery's pattern.
- */
-function createTaskStatusQuery(opts?: {
-  fetchFn?: typeof fetch;
-}): (repo: string, prNumber: number) => Promise<string | null> {
-  const taskStoreUrl = (process.env.SHIPWRIGHT_TASK_STORE_URL ?? "").trim();
-  const taskStoreToken = (process.env.SHIPWRIGHT_TASK_STORE_TOKEN ?? "").trim();
-  const doFetch = opts?.fetchFn ?? fetch;
-
-  return async (repo: string, prNumber: number): Promise<string | null> => {
-    if (!taskStoreUrl || !taskStoreToken) return null;
-    try {
-      const baseUrl = taskStoreUrl.replace(/\/$/, "");
-      const params = new URLSearchParams({ repo, pr: String(prNumber) });
-      const res = await doFetch(`${baseUrl}/tasks?${params}`, {
-        headers: {
-          Authorization: `Bearer ${taskStoreToken}`,
-          "Content-Type": "application/json",
-        },
-      });
-      if (!res.ok) return null;
-      const data = (await res.json()) as unknown;
-      let tasks: Array<{ status: string }> = [];
-      if (Array.isArray(data)) {
-        tasks = data as Array<{ status: string }>;
-      } else if (
-        data !== null &&
-        typeof data === "object" &&
-        Array.isArray((data as Record<string, unknown>).tasks)
-      ) {
-        tasks = (data as Record<string, unknown>).tasks as Array<{
-          status: string;
-        }>;
-      }
-      return tasks[0]?.status ?? null;
-    } catch {
-      return null;
-    }
   };
 }
